+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [unreleased]
-There's Nothing but Crickets ¯\\_(ツ)_/¯
+- Fixed: Telegram invalid token error
+- Added to automatically check whether the docker image is working properly or not
 
 ## [0.1.6-alpha] - 2023-12-01
 - Added docker image testing in Github Actions
@@ -44,5 +45,5 @@ There's Nothing but Crickets ¯\\_(ツ)_/¯
 [0.1.3-alpha]: https://github.com/BiltuDas1/vigilantbot-telegram/compare/0.1.2...0.1.3
 [0.1.4-alpha]: https://github.com/BiltuDas1/vigilantbot-telegram/compare/0.1.3...0.1.4
 [0.1.5-alpha]: https://github.com/BiltuDas1/vigilantbot-telegram/compare/0.1.4...0.1.5
-[0.1.6-alpha] - 2023-12-01: https://github.com/BiltuDas1/vigilantbot-telegram/compare/0.1.5...main
+[0.1.6-alpha]: https://github.com/BiltuDas1/vigilantbot-telegram/compare/0.1.5...0.1.6
 [unreleased]: https://github.com/BiltuDas1/vigilantbot-telegram/compare/0.1.6...main